add correct minimum slider index
